+++ src/tools/clippy/tests/ui-toml/disallowed_names_append/disallowed_names.rs
@@ -1,9 +1,9 @@
-#[warn(clippy::blacklisted_name)]
+#[warn(clippy::disallowed_names)]
 
 fn main() {
     // `foo` is part of the default configuration
     let foo = "bar";
-    // `ducks` was unrightfully blacklisted
+    // `ducks` was unrightfully disallowed
     let ducks = ["quack", "quack"];
     // `fox` is okay
     let fox = ["what", "does", "the", "fox", "say", "?"];